fix(native): no more shame
I just found a way of incorporating handlers without passing "onFoo" names in props, and it turns
out to be pretty simple. This commit removes the ReactNativeEventEmitter-based lower level sources,
and now only has the handler-based architecture, removing the name shameful because we are no longer
passing onFoo props, only the selector prop is needed. This implementations assume intent() will be
called before view() is called (so that the handler subjects already exist when the view is called,
so the view through hyperscript can simply attach those existing ones). Using hyperscript here is
essential to be able to tap into the ReactElement creation.